Migrate Day1 from u64 to usize

+++ 2025/day1/src/lib.rs
@@ -1,11 +1,11 @@
-pub fn solve_task_one(file: &str) -> u64 {
+pub fn solve_task_one(file: &str) -> usize {
     let mut list_one = vec![];
     let mut list_two = vec![];
 
     for line in file.trim().lines() {
         let mut line_parts = line.split_ascii_whitespace();
-        list_one.push(line_parts.next().unwrap().parse::<u64>().unwrap());
-        list_two.push(line_parts.next().unwrap().parse::<u64>().unwrap());
+        list_one.push(line_parts.next().unwrap().parse::<usize>().unwrap());
+        list_two.push(line_parts.next().unwrap().parse::<usize>().unwrap());
     }
 
     list_one.sort();
@@ -15,17 +15,17 @@ pub fn solve_task_one(file: &str) -> u64 {
         .iter()
         .zip(list_two.iter())
         .map(|(one, two)| one.abs_diff(*two))
-        .sum::<u64>()
+        .sum::<usize>()
 }
 
-pub fn solve_task_two(file: &str) -> u64 {
+pub fn solve_task_two(file: &str) -> usize {
     let mut list_one = vec![];
     let mut list_two = vec![];
 
     for line in file.trim().lines() {
         let mut line_parts = line.split_ascii_whitespace();
-        list_one.push(line_parts.next().unwrap().parse::<u64>().unwrap());
-        list_two.push(line_parts.next().unwrap().parse::<u64>().unwrap());
+        list_one.push(line_parts.next().unwrap().parse::<usize>().unwrap());
+        list_two.push(line_parts.next().unwrap().parse::<usize>().unwrap());
     }
 
     list_one
@@ -35,7 +35,7 @@ pub fn solve_task_two(file: &str) -> u64 {
                 * list_two
                     .iter()
                     .filter(|list_two_item| list_one_item == *list_two_item)
-                    .count() as u64
+                    .count() as usize
         })
-        .sum::<u64>()
+        .sum::<usize>()
 }
